Tighter References
References are now tighter, meaning they are only used for accessing SRoots from a SObject.
