+++ lib/serekit/sobjects/sobjects.dart
@@ -2,4 +2,3 @@ export 'description.dart';
 export 'header.dart';
 export 'file_system.dart';
 export 'settings.dart';
-export 'indent.dart';
