+++ lib/version_control/star.dart
@@ -66,8 +66,7 @@ class Star extends SObject {
     await getParent<Star>()!.makeCurrent();
     await archive.then((e) => e!.markForDeletion());
     for (final archiveReference in getDescendants<SRArchive>()) {
-      kit.addDeletionRequest(
-          SDeletionRequest<SArchive>(archiveReference!.hash));
+      archiveReference!.markForDeletion();
     }
     unparent();
   }
